requirements.txt
numpy
onnxruntime-gpu
onnxruntime-directml
fastapi
uvicorn[standard]